Move build code closer to where it's used
