+++ buildlib.py
@@ -5,8 +5,6 @@ import sys
 import subprocess
 import traceback
 
-import patch_ng as patch
-
 
 def is_mac():
 	return platform.system() == 'Darwin'
@@ -56,10 +54,6 @@ class Config:
 	file_aq_release = os.path.join(base, 'audioquake', 'release')
 
 
-#
-# Utilities
-#
-
 def comeback(function):
 	def wrapper(*args, **kwargs):
 		original = os.getcwd()
@@ -105,18 +99,6 @@ def try_to_run(process_args, error_message):
 			die(error_message)
 
 
-#
-# Engine compilation
-#
-
-def compile_zqcc():
-	_compile(Config.dir_make_zqcc, 'zqcc')
-
-
-def compile_zquake():
-	_compile(Config.dir_make_zquake, 'zquake', ['gl', 'server'])
-
-
 def _make(name, target=None):
 	error_message = 'failed to compile ' + name + ' [target: ' + str(target) + ']'
 	process_args = ['make']
@@ -126,7 +108,7 @@ def _make(name, target=None):
 
 
 @comeback
-def _compile(path, name, targets=[]):
+def make(path, name, targets=[]):
 	try:
 		os.chdir(path)
 	except:  # noqa E727
@@ -136,114 +118,3 @@ def _compile(path, name, targets=[]):
 	else:
 		for targ in targets:
 			_make(name, targ)
-
-
-def compile_zquake_windows():
-	path = Config.dir_zquake_source
-	try:
-		os.chdir(path)
-	except:  # noqa E727
-		die("can't change directory to: " + path)
-	try_to_run(
-		['msbuild', 'zquake.sln', '/p:Configuration=GLRelease', '/p:Platform=Win32'],
-		'ZQuake compilation')
-
-
-def compile_zqcc_windows():  # FIXME DRY
-	path = Config.dir_make_zqcc
-	try:
-		os.chdir(path)
-	except:  # noqa E727
-		die("can't change directory to: " + path)
-	try_to_run(
-		['msbuild', 'zqcc.sln', '/p:Configuration=Release', '/p:Platform=Win32'],
-		'ZQCC compilation')
-
-
-#
-# QuakeC Compilation
-#
-
-def _chdir_gamecode():  # TODO remove
-	try:
-		os.chdir(Config.dir_qc)
-	except:  # noqa E727
-		die("can't change to QuakeC directory: " + Config.dir_qc)
-
-
-@comeback
-def compile_gamecode():
-	_chdir_gamecode()
-	_compile_gamecode('progs.src')
-	_compile_gamecode('spprogs.src')
-
-
-def _compile_gamecode(progs):
-	try_to_run(
-		(os.path.join(Config.dir_qc, Config.bin_zqcc), '-progs', progs),
-		'failed to compile gamecode file: ' + progs)
-
-
-#
-# Map tools compilation
-#
-
-def rename_qutils():
-	for root, files, dirs in os.walk(Config.dir_qutils, topdown=False):
-		for name in files + dirs:
-			os.rename(
-				os.path.join(root, name),
-				os.path.join(root, name.lower()))
-
-
-def patch_map_tools():
-	patches = {
-		'Makefile': os.path.join(Config.dir_patches, 'makefile.patch'),
-		'writebsp.c': os.path.join(Config.dir_patches, 'writebsp.c.patch'),
-		'qbsp.c': os.path.join(Config.dir_patches, 'qbsp.c.patch')
-	}
-
-	for title, patch_file in patches.items():
-		print('Patching', title)
-		patch_set = patch.fromfile(patch_file)
-		if not patch_set.apply(root=Config.dir_qbsp):
-			raise Exception('Patch', patch_file, 'failed.')
-
-	if is_windows():
-		patch_map_tools_windows()
-
-
-def patch_map_tools_windows():
-	windows_patches = {
-		'qbsp.mak': os.path.join(Config.dir_patches, 'qbsp.mak.patch'),
-		'light.mak': os.path.join(Config.dir_patches, 'light.mak.patch'),
-		'vis.mak': os.path.join(Config.dir_patches, 'vis.mak.patch'),
-		'bspinfo.mak': os.path.join(Config.dir_patches, 'bspinfo.mak.patch')
-	}
-
-	# FIXME DRY
-	for title, patch_file in windows_patches.items():
-		print('Patching', title)
-		patch_set = patch.fromfile(patch_file)
-		if not patch_set.apply(root=Config.dir_quake_tools):
-			raise Exception('Patch', patch_file, 'failed.')
-
-
-def compile_map_tools():
-	_compile(Config.dir_qbsp, 'Quake map tools')
-
-
-@comeback
-def compile_map_tools_windows():  # FIXME DRY
-		for prog in ['qbsp', 'vis', 'light', 'bspinfo']:
-			print('Map tool:', prog)
-			path = os.path.join(Config.dir_qutils, prog)
-			try:
-					os.chdir(path)
-			except:  # noqa E727
-					die("can't change directory to: " + path)
-			try_to_run([
-				'nmake',
-				'/f', prog + '.mak',
-				'CFG=' + prog + ' - Win32 Release'],
-				prog + ' compilation')
